Extract embedded benchmark scripts into standalone files
